+++ internal/crossdock/client/dispatcher/dispatcher.go
@@ -29,9 +29,9 @@ import (
 	"go.uber.org/yarpc/encoding/raw"
 	"go.uber.org/yarpc/internal/crossdock/client/params"
 	"go.uber.org/yarpc/internal/yarpctest"
+	"go.uber.org/yarpc/transport/grpc"
 	"go.uber.org/yarpc/transport/http"
 	"go.uber.org/yarpc/transport/tchannel"
-	"go.uber.org/yarpc/transport/x/grpc"
 )
 
 // Create creates an RPC from the given parameters or fails the whole behavior.
